move transmittal sheet common db code to common
