+++ hmda-analytics/src/main/scala/hmda/analytics/HmdaAnalyticsApp.scala
@@ -9,11 +9,11 @@ import akka.stream.ActorMaterializer
 import akka.stream.scaladsl.{Keep, Sink, Source}
 import akka.util.Timeout
 import com.typesafe.config.ConfigFactory
+import hmda.query.ts.TransmittalSheetConverter
 import hmda.analytics.query.{
   LarComponent,
   LarConverter,
-  TransmittalSheetComponent,
-  TransmittalSheetConverter
+  TransmittalSheetComponent
 }
 import hmda.messages.pubsub.HmdaTopics.{analyticsTopic, signTopic}
 import hmda.model.filing.lar.LoanApplicationRegister
