+++ hmda-regulator/src/main/scala/hmda/regulator/query/RegulatorComponent.scala
@@ -4,7 +4,7 @@ import hmda.query.DbConfiguration._
 import hmda.query.repository.TableRepository
 import hmda.regulator.query.lar.{LarEntityImpl, _}
 import hmda.regulator.query.panel.{InstitutionEmailEntity, InstitutionEntity}
-import hmda.regulator.query.ts.TransmittalSheetEntity
+import hmda.query.ts.TransmittalSheetEntity
 import slick.basic.DatabaseConfig
 import slick.jdbc.JdbcProfile
